Topicの依存を解消
