+++ app/build.gradle
@@ -70,6 +70,7 @@ android {
 }
 
 dependencies {
+    implementation project(":core:model")
 
     implementation 'androidx.core:core-ktx:1.8.0'
     implementation "androidx.compose.ui:ui:1.2.1"
